+++ RELEASE-NOTES.md
@@ -18,6 +18,7 @@
 * #478 ifNull($col(...), $col(...)) does not compile
 * #483 Upgrade ECharts to 5.6.0
 * #484 Upgrade Parquet and Hadoop
+* #487 Series.compact()
 
 ## Release 2.0.0-M2